pyproject.toml
[tool.poetry]
name = "sky-pro-pd-hw27-hw32"
version = "0.1.0"
description = ""
authors = ["grm <golosov_sa@mail.ru>"]
readme = "README.md"
packages = [{include = "sky"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.2"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"
pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
